+++ dbtools-room/src/commonMain/kotlin/org/dbtools/room/converter/KotlinDateTimeTextConverter.kt
@@ -1,6 +1,6 @@
 package org.dbtools.room.converter
 
-import androidx.room.TypeConverter
+import androidx.room3.TypeConverter
 import kotlinx.datetime.LocalDate
 import kotlinx.datetime.LocalDateTime
 import kotlinx.datetime.LocalTime